PyLearning/requirements.txt
colorgram.py==1.2.0
pillow==12.1.0
prettytable==3.17.0
wcwidth==0.4.0
